Fixed Interaction Selection
It now chooses the item that is closest to the playerHand(GameObject/child).

Only need to consider the distance from the interactable object and the InteractArrow(highlight/selection icon) - should it be higher, like above the player to be more visible, or simply just on top of the interactable object? Maybe even have it adapt to when the player is covering the object to then sit above?