Player Pause when meeting Waldu first time
The HaltPlayer script disables the playerController script when entering the trigger of Waldu, then when text is hiding, player regains control again - this happends only on first meeting.

However, need more opinions on how to handle the holding of the player: Should the player freeze in its animation (this gives the most control over the player) or should the player still be playing the Idle animation to still look alive when the message is conveyed? The latter makes it though possible to land in fron of the Waldu, risking that the player will miss the cure character's introduction.

Also, would need to add the Vignette to help the focus! and possibly a slightly longer pause when triggering Waldu.