Applied suggested changes
lowered the volume and bass of the running sound for the wolf, this was amplified on my headphones, which are normally set pretty low, so didn't hear until I set them a bit higher! We'll see if this works otherwise we'll have to look for something else..

Also, I readjusted the timing of the steps to be a better fit with the wolf's running animation, please note this.